YYC-111: split memory.rs into schema + codec submodules
memory.rs covered the SQL schema constant + idempotent migration
helpers, the message encode/decode helpers, the SessionStore CRUD
surface, and the test suite — ~966 lines in one file.

Split into three focused modules:

* memory/schema.rs — SCHEMA DDL constant, initialize_conn,
  upsert_session_metadata, upsert_session_provider_profile,
  open_gateway_connection (gateway-only).
* memory/codec.rs — Encoded type alias, encode_message,
  decode_message.
* memory/mod.rs — SessionStore, SearchHit, SessionSummary, impl
  block, Default impl, integration tests.

Schema/codec items are pub(in crate::memory) so SessionStore can
call them while keeping them invisible outside the memory module.
The gateway-feature exports (open_gateway_connection,
initialize_test_conn) are re-exported via pub(crate) use so call
sites in src/gateway/ continue to work unchanged.

Public API unchanged. cargo test --all-features --lib passes
(250 tests).
